move functionality into its own component in Connected
