+++ src/app/components/blockchain/Connected/Connected.tsx
@@ -1,13 +1,16 @@
-import { Box, NetworkImage, Text } from "@0xsequence/design-system";
+import { Box, Text } from "@0xsequence/design-system";
 import { useAccount, useDisconnect } from "wagmi";
+import ActiveNetwork from "../ActiveNetwork";
+import ChainEnvironment from "../ChainEnvironment";
+import TestSendTransaction from "../../Transactions";
+import NativeBalance from "../NativeBalance";
 
 const Connected = () => {
-  const { address } = useAccount();
+  const { address, chain } = useAccount();
   const { disconnect } = useDisconnect();
   const onClickDisconnect = () => {
     disconnect();
   };
-
   return (
     <>
       <Text variant="large" fontWeight="bold" color="text100">
@@ -16,6 +19,16 @@ const Connected = () => {
       <div className="card">
         <button onClick={onClickDisconnect}>Disconnect</button>
       </div>
+      {chain && (
+        <Box marginBottom="8">
+          <Box display="flex" justifyContent="space-between">
+            <ActiveNetwork />
+            <ChainEnvironment />
+          </Box>
+          <NativeBalance />
+        </Box>
+      )}
+      <TestSendTransaction />
     </>
   );
 };
